robot_gateway/pyproject.toml
[project]
name = "robot-gateway"
version = "0.1.0"
description = "OpenClaw data-plane robot gateway"
readme = "references/robot-gateway-mvp.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn>=0.29.0",
  "pydantic>=2.6.0",
  "httpx>=0.26.0",
  "pillow>=10.0.0",
  "aiortc>=1.7.0",
  "av>=11.0.0",
  "opencv-python>=4.8.0",
  "paho-mqtt>=1.6.1",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
]

[tool.uv]
package = false
